requirements.txt
# === Django Framework ===
django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1

# === External Services ===
twilio==8.10.0
openai==1.97.0

# === Environment and Configuration ===
python-dotenv==1.0.0
gunicorn==21.2.0

# === Voice Processing (Optional) ===
pyttsx3==2.90
SpeechRecognition==3.10.0
# pyaudio==0.2.11  # Commented due to PortAudio system dependency

# === System Monitoring ===
psutil==5.9.6

# === PostgreSQL Support ===
psycopg2-binary==2.9.9

# === Caching and Session Management ===
redis==5.0.1

# === HTTP Requests and Integrations ===
requests==2.31.0
python-dateutil==2.8.2

# === Legacy Flask Dependencies (Removed) ===
# flask==3.0.0 - Replaced with Django
# sqlalchemy==2.0.23 - Using Django ORM
# Jinja2==3.1.2 - Using Django templates
# MarkupSafe==2.1.3 - Built into Django
# click==8.1.7 - Not needed with Django
# itsdangerous==2.1.2 - Built into Django
# Werkzeug==3.0.1 - Not needed with Django